+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Improved build pipeline to use [codaamok.build](https://github.com/codaamok/codaamok.build)
+- New `-ForwardQuery` boolean parameter in `New-ShlinkUrl` and `Set-ShlinkUrl` (new in Shlink 2.9.0).
 
 ### Changed
+- Minimum Shlink version updated to 2.9.0
+- Renamed parameter `-DoNotValidateUrl` of switch type to `-ValidateUrl` of boolean type for both `New-ShlinkUrl` and `Set-ShlinkUrl`.
+- Parameter `-FindIfExists` is now of type boolean instead of switch in `New-ShlinkUrl`. is now a switch and not a boolean`.
 - Removed use of methods `SecureStringToBSTR` and `PtrToStringAuto` from Marshal .NET class for secure string handling in `InvokeShlinkRestMethod`, in favour of using `GetNetworkCredential` method from a PSCredential object.
 - `Save-ShlinkUrlQrCode` no longer has hardcoded default parameter values (much like the API did) for size, margin, format, and error correction level.. Shlink 2.9.0 now lets you configure these defaults for QR codes, so by omitting values for these params, your server default values are used.